make compile again for 6.1.0 & 3.5.0 + fix some warnings

+++ src/Common.h
@@ -8,7 +8,7 @@ constexpr char stringOuterDelimiter[] = "^"; // Character used to encapsulate en
 
 inline bool isNumber(const char *str)
 {
-	byte i = 0;
+	int i = 0;
 
 	while (*(str + i) != '\0') {
 		if (!isdigit(*(str + i++))) {
